requirements.txt
streamlit
pandas
sqlalchemy
pymysql
langchain
langchain-community
langchain-google-genai
weasyprint
xlsxwriter
python-dotenv